# doc/api.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
  /user:
    post:
      operationId: UserController_create
      summary: Create user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: User created successfully
        '400':
          description: Bad request. Body does not contain required fields
      tags:
        - User
    get:
      operationId: UserController_findAll
      summary: Get all users
      parameters: []
      responses:
        '200':
          description: Return all users
      tags:
        - User
  /user/{id}:
    get:
      operationId: UserController_findOne
      summary: Get user by id
      parameters:
        - name: id
          required: true
          in: path
          description: User UUID
          schema:
            type: string
      responses:
        '200':
          description: Return user by id
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User not found
      tags:
        - User
    put:
      operationId: UserController_update
      summary: Update user password
      parameters:
        - name: id
          required: true
          in: path
          description: User UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPasswordDto'
      responses:
        '200':
          description: User password updated successfully
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User not found
      tags:
        - User
    delete:
      operationId: UserController_remove
      summary: Delete user
      parameters:
        - name: id
          required: true
          in: path
          description: User UUID
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User not found
      tags:
        - User
  /artist:
    post:
      operationId: ArtistController_create
      summary: Create artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: Artist created successfully
        '400':
          description: Bad request. Body does not contain required fields
      tags:
        - Artist
    get:
      operationId: ArtistController_findAll
      summary: Get all artists
      parameters: []
      responses:
        '200':
          description: Return all artists
      tags:
        - Artist
  /artist/{id}:
    get:
      operationId: ArtistController_findOne
      summary: Get artist by id
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            type: string
      responses:
        '200':
          description: Return artist by id
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Artist
    put:
      operationId: ArtistController_update
      summary: Update artist information
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistInfoDto'
      responses:
        '200':
          description: Artist updated successfully
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Artist
    delete:
      operationId: ArtistController_remove
      summary: Delete artist
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            type: string
      responses:
        '204':
          description: Artist deleted successfully
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Artist
  /track:
    post:
      operationId: TrackController_create
      summary: Create track
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: Track created successfully
        '400':
          description: Bad request. Body does not contain required fields
      tags:
        - Track
    get:
      operationId: TrackController_findAll
      summary: Get all tracks
      parameters: []
      responses:
        '200':
          description: Return all tracks
      tags:
        - Track
  /track/{id}:
    get:
      operationId: TrackController_findOne
      summary: Get track by id
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            type: string
      responses:
        '200':
          description: Return track by id
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags:
        - Track
    put:
      operationId: TrackController_update
      summary: Update track information
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: Track updated successfully
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags:
        - Track
    delete:
      operationId: TrackController_remove
      summary: Delete track
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            type: string
      responses:
        '204':
          description: Track deleted successfully
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags:
        - Track
  /album:
    post:
      operationId: AlbumController_create
      summary: Create album
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: Album created successfully
        '400':
          description: Bad request. Body does not contain required fields
      tags:
        - Album
    get:
      operationId: AlbumController_findAll
      summary: Get all albums
      parameters: []
      responses:
        '200':
          description: Return all albums
      tags:
        - Album
  /album/{id}:
    get:
      operationId: AlbumController_findOne
      summary: Get album by id
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            type: string
      responses:
        '200':
          description: Return album by id
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Album
    put:
      operationId: AlbumController_update
      summary: Update album information
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: Album updated successfully
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Album
    delete:
      operationId: AlbumController_remove
      summary: Delete album
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            type: string
      responses:
        '204':
          description: Album deleted successfully
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Album
  /favs:
    get:
      operationId: FavsController_getFavorites
      summary: Get all favorites
      parameters: []
      responses:
        '200':
          description: Get all favorites movies, tracks and books
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritesResponseDto'
      tags:
        - Favorites
  /favs/track/{id}:
    post:
      operationId: FavsController_addTrackToFavorites
      summary: Add track to favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            type: string
      responses:
        '201':
          description: Track added to favorites successfully
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '422':
          description: Track with id doesn't exist
      tags:
        - Favorites
    delete:
      operationId: FavsController_removeTrackFromFavorites
      summary: Delete track from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Track UUID
          schema:
            type: string
      responses:
        '204':
          description: Track deleted from favorites successfully
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track was not found in favorites
      tags:
        - Favorites
  /favs/album/{id}:
    post:
      operationId: FavsController_addAlbumToFavorites
      summary: Add album to favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            type: string
      responses:
        '201':
          description: Album added to favorites successfully
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '422':
          description: Album with id doesn't exist
      tags:
        - Favorites
    delete:
      operationId: FavsController_removeAlbumFromFavorites
      summary: Delete album from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Album UUID
          schema:
            type: string
      responses:
        '204':
          description: Album deleted from favorites successfully
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album was not found in favorites
      tags:
        - Favorites
  /favs/artist/{id}:
    post:
      operationId: FavsController_addArtistToFavorites
      summary: Add artist to favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            type: string
      responses:
        '201':
          description: Artist added to favorites successfully
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '422':
          description: Artist with id doesn't exist
      tags:
        - Favorites
    delete:
      operationId: FavsController_removeArtistFromFavorites
      summary: Delete artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: Artist UUID
          schema:
            type: string
      responses:
        '204':
          description: Artist deleted from favorites successfully
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist was not found in favorites
      tags:
        - Favorites
info:
  title: Home Library API
  description: The Home Library API description
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          description: User login
          example: User
        password:
          type: string
          description: User password
          example: '1488'
      required:
        - login
        - password
    UpdateUserPasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          description: Current password
          example: '1488'
        newPassword:
          type: string
          description: New password
          example: '1234'
      required:
        - oldPassword
        - newPassword
    CreateArtistDto:
      type: object
      properties: {}
    UpdateArtistInfoDto:
      type: object
      properties:
        name:
          type: string
          example: Eminem
          description: Updated artist name
        grammy:
          type: boolean
          example: true
          description: Updated Grammy award status
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: Mamma Maria
          description: Track name
        artistId:
          type: string
          example: 390cbe23-e4c9-43e6-828d-7795833cad90
          description: Optional artist ID (UUID v4)
          nullable: true
        albumId:
          type: string
          example: 390cbe23-e4c9-43e6-828d-7795833cad90
          description: Optional album ID (UUID v4)
          nullable: true
        duration:
          type: number
          example: 200
          description: Duration in seconds (integer > 0)
      required:
        - name
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
        duration:
          type: number
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          example: Vol1
          description: Album name
        year:
          type: number
          example: 2020
          description: Release year
        artistId:
          type: string
          example: 1a848183-5265-4af1-9a82-048695e31b0b
          description: Optional artist ID (UUID v4)
          nullable: true
      required:
        - name
        - year
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          example: Vol2
          description: Updated album name
        year:
          type: number
          example: 2021
          description: Updated release year
        artistId:
          type: string
          example: null
          description: Artist ID (set to null to remove association)
          nullable: true
    Artist:
      type: object
      properties: {}
    Album:
      type: object
      properties: {}
    Track:
      type: object
      properties: {}
    FavoritesResponseDto:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
      required:
        - artists
        - albums
        - tracks